refactor SubscriptionBehavior into Acceptance Testing framework
