+++ src/NServiceBus.AcceptanceTests/NServiceBus.AcceptanceTests.csproj
@@ -227,7 +227,6 @@
     <Compile Include="Core\AutomaticSubscriptions\When_starting_an_endpoint_with_autosubscribe.cs">
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="Routing\SubscriptionBehaviorExtensions.cs" />
     <Compile Include="Sagas\When_saga_handles_unmapped_message.cs" />
     <Compile Include="Satellites\When_a_message_is_available.cs" />
     <Compile Include="Serialization\When_registering_deserializers_with_settings.cs" />
@@ -298,11 +297,9 @@
     <Compile Include="Outbox\When_receiving_a_message.cs" />
     <Compile Include="NServiceBusAcceptanceTest.cs" />
     <Compile Include="DataBus\When_sending_databus_properties.cs" />
-    <Compile Include="Routing\SubscriptionBehavior.cs" />
     <Compile Include="Routing\MessageDrivenSubscriptions\When_multi_subscribing_to_a_polymorphic_event.cs" />
     <Compile Include="Routing\When_publishing_with_overridden_local_address.cs" />
     <Compile Include="Routing\When_publishing_with_only_local_messagehandlers.cs" />
-    <Compile Include="Routing\SubscriptionEventArgs.cs" />
     <Compile Include="Routing\When_base_event_from_2_publishers.cs" />
     <Compile Include="Routing\When_publishing.cs" />
     <Compile Include="Recoverability\When_immediate_retries_fail.cs" />
